# pages/api/search/jadwalsholat.js
import fetch from 'node-fetch';
import { trackRequest } from "../../../lib/redis";
import { API_KEY, CREATOR } from '../../../settings';

export default async function handler(req, res) {
    if (req.method !== 'GET') {
        return res.status(405).json({
            status: false,
            creator: CREATOR,
            error: 'Method Not Allowed',
        });
    }

    const { query } = req.query;

    if (!query) {
        return res.status(400).json({
            status: false, 
            creator: CREATOR, 
            error: 'missing query parameter'
        });
    }
    
    try {
        await trackRequest("/api/search/jadwalsholat");
        const result = await jadwalsholat(query);
        res.status(200).json({
            status: true,
            creator: CREATOR,
            data: result,
        });
    } catch (error) {
        res.status(500).json({
            status: false,
            creator: CREATOR,
            error: 'Internal Server Error',
        });
    }
}

async function jadwalsholat(query) {
  const response = await fetch("https://jadwalsholat-eight.vercel.app/api/jadwalsholat?text=" + query);
  const result = await response.json();

  return result.result;
}