# pages/swagger-config.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Xavier Api",
    "description": "Welcome to the Xavier Api, your gateway to streamlined and efficient application integration. Designed with simplicity and versatility in mind, Xavier Api empowers developers to effortlessly manage and interact with essential data and functionalities. Whether you're building a web application, mobile app, or integrating with other services, Laurine Api provides a robust foundation for seamless operations.",
    "contact": {
      "email": "https://t.me/FVendettaV"
    },
    "license": {
      "name": "API Development Information",
      "url": "https://whatsapp.com/channel/0029VbACugX6rsQyHw0Wvr1g"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://japi.vercel.app",
      "description": "Xavier Api"
    }
  ],
  "paths": {
...
    "/api/orkut/payment": {
      "get": {
        "tags": ["OrderKuota"],
        "description": "This endpoint is used to start the payment process with a nominal amount adjusted for the OrderKuota user's order.This endpoint is used to start the payment process with a nominal amount adjusted for the OrderKuota user's order.",
        "parameters": [
          {
            "name": "amount",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "codeqr",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiKey",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/downloader/ytmp3": {
      "get": {
        "tags": ["Downloader"],
        "description": "YTDLMP3 stands for YouTube Downloader, a tool used to download videos from YouTube and other websites. Popular implementations include youtube-dl and yt-dlp, which are command-line based and open-source.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
              }
           }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/downloader/threadsdl": {
      "get": {
        "tags": ["Downloader"],
        "description": "Threads Downloader is a tool or application that allows users to download videos, images, or other content from the Threads social media platform. Threads is a text and image-based app developed by Meta (Instagram) for sharing thoughts, discussions, and visual media. This tool is commonly used to save or repost content from Threads without needing to take screenshots or screen recordings.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
              }
           }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/generator/qc": {
      "get": {
        "tags": ["Generator"],
        "description": "QC (Quote Creator) is a feature that converts text quotes into images with a stylish design. It typically includes elements like the author's name, custom fonts, and backgrounds to enhance the visual appeal.",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {            
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {        
            "name": "photo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/search/tiktoksearch": {
      "get": {
        "tags": ["Search"],
        "description": "TikTok Search is a video search service for TikTok that can be accessed through various platforms, including TikWM (tikwm.com). This site provides an API that allows users to search for TikTok videos by keywords and retrieve download links for videos with or without watermarks, music, covers, and creator avatars.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/search/rule34search": {
      "get": {
        "tags": ["Search"],
        "description": "Rule34 is an internet rule that means if an anime or manga character exists, there is NSFW or adult content of it. This concept is popular in fandom communities and is often associated with fan art, doujinshi, or other fan-created works in the form of illustrations and media.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/search/jadwalsholat": {
      "get": {
        "tags": ["Search"],
        "description": "Jadwalsholat is a web-based service that provides prayer time schedules for various regions in Indonesia. Through this website, users can view daily, weekly, or monthly prayer times based on their selected location. Additionally, some prayer schedule services offer extra features such as prayer reminders, Qibla direction, and an Islamic calendar to help Muslims perform their religious duties more consistently.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}